+++ vkconfig_core/test/test_configuration.cpp
@@ -27,6 +27,14 @@
 
 #include <gtest/gtest.h>
 
+static void InitLayer(LayerManager& layer_manager) {
+    const std::vector<Path>& layers_paths = ::CollectFilePaths(":/layers");
+
+    for (std::size_t i = 0, n = layers_paths.size(); i < n; ++i) {
+        layer_manager.LoadLayer(layers_paths[i], LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    }
+}
+
 static bool operator==(const Configuration& a, const Configuration& b) {
     if (a.key != b.key)
         return false;
@@ -194,7 +202,7 @@ TEST(test_configuration, make_duplicate_tagged_name_mix) {
 
 TEST(test_configuration, create) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
 
@@ -214,7 +222,7 @@ TEST(test_configuration, create) {
 
 TEST(test_configuration, create_disabled) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::CreateDisabled(layer_manager);
 
@@ -234,7 +242,7 @@ TEST(test_configuration, create_disabled) {
 
 TEST(test_configuration, SwitchLayerVersion) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
     for (std::size_t i = 0, n = configuration.parameters.size(); i < n; ++i) {
@@ -261,7 +269,7 @@ TEST(test_configuration, SwitchLayerVersion) {
 
 TEST(test_configuration, gather_parameters_exist) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
@@ -294,7 +302,7 @@ TEST(test_configuration, gather_parameters_exist) {
 
 TEST(test_configuration, gather_parameters_repeat) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
@@ -324,11 +332,11 @@ TEST(test_configuration, gather_parameters_repeat) {
 
 TEST(test_configuration, gather_parameters_missing) {
     LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
 
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
     configuration.GatherParameters(layer_manager);
 
     layer_manager.Clear();
@@ -361,11 +369,11 @@ TEST(test_configuration, gather_parameters_missing) {
 
 TEST(test_configuration, HasMissingLayer_UnsupportPlatform) {
     LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
 
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
     configuration.GatherParameters(layer_manager);
 
     layer_manager.Clear();
@@ -381,11 +389,11 @@ TEST(test_configuration, HasMissingLayer_UnsupportPlatform) {
 
 TEST(test_configuration, HasMissingLayer_Off) {
     LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
 
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
     configuration.GatherParameters(layer_manager);
 
     layer_manager.Clear();
@@ -401,11 +409,11 @@ TEST(test_configuration, HasMissingLayer_Off) {
 
 TEST(test_configuration, gather_parameters_missing_but_unsupported_platform) {
     LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
     configuration.key = "New Configuration";
 
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
     configuration.GatherParameters(layer_manager);
 
     layer_manager.Clear();
@@ -421,7 +429,7 @@ TEST(test_configuration, gather_parameters_missing_but_unsupported_platform) {
 
 TEST(test_configuration, Reorder_full) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
 
@@ -457,7 +465,7 @@ TEST(test_configuration, Reorder_full) {
 
 TEST(test_configuration, Reorder_partial) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
 
@@ -485,7 +493,7 @@ TEST(test_configuration, Reorder_partial) {
 
 TEST(test_configuration, Reorder_missing) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
 
@@ -515,7 +523,7 @@ TEST(test_configuration, Reorder_missing) {
 
 TEST(test_configuration, Reorder_empty) {
     LayerManager layer_manager;
-    layer_manager.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    ::InitLayer(layer_manager);
 
     Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
     std::size_t size = configuration.parameters.size();
@@ -527,19 +535,19 @@ TEST(test_configuration, Reorder_empty) {
 }
 
 TEST(test_configuration, IsDefault_True) {
-    LayerManager layers;
-    layers.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
     Configuration configuration;
-    const bool loaded = configuration.Load(":/configurations/Validation.json", layers);
+    const bool loaded = configuration.Load(":/configurations/Validation.json", layer_manager);
     EXPECT_TRUE(loaded);
     EXPECT_TRUE(configuration.IsDefault());
 }
 
 TEST(test_configuration, IsDefault_False) {
-    LayerManager layers;
-    layers.LoadLayersFromPath(":/layers", LAYER_TYPE_EXPLICIT, CONFIGURATOR_MODE_CMD);
+    LayerManager layer_manager;
+    ::InitLayer(layer_manager);
 
-    Configuration configuration = Configuration::Create(layers, "New Configuration");
+    Configuration configuration = Configuration::Create(layer_manager, "New Configuration");
     EXPECT_FALSE(configuration.IsDefault());
 }