+++ vkconfig_core/type_layers_paths.h
@@ -39,3 +39,5 @@ enum LayersPaths {
 enum { LAYERS_PATHS_COUNT = LAYERS_PATHS_LAST - LAYERS_PATHS_FIRST + 1 };
 
 const char* GetLabel(LayersPaths Layers_paths_type);
+
+LayerType GetLayerType(LayersPaths Layers_paths_type);